+++ thefesta/src/board/css/BoardList.css
@@ -1,5 +1,23 @@
 .board-container {
-    max-width: 1200px;
+    min-width: 1024px;
+    margin: 0 auto;
+    padding: 20px;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  }
+  .board-item {
+    min-width: 768px;
+    max-width: 968px;
+    margin: 0 auto;
+    padding: 20px;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  }
+  .reply-item {
+    min-width: 768px;
+    max-width: 968px;
     margin: 0 auto;
     padding: 20px;
     border: 1px solid #ddd;
