+++ thefesta/src/board/css/BoardRegister.css
@@ -1,6 +1,6 @@
 /* CSS 코드 */
 .board-register-container {
-    max-width: 600px;
+    max-width: 800px;
     margin: 0 auto;
     padding: 20px;
     border: 1px solid #ddd;
@@ -27,15 +27,15 @@
   input,
   textarea {
     width: 100%;
-    padding: 10px;
+    min-height: auto;
     font-size: 16px;
     border: 1px solid #ddd;
     border-radius: 4px;
   }
   
   button {
-    background-color: #007bff;
-    color: #fff;
+    background-color: #ffffff;
+    color: #030303;
     padding: 10px 20px;
     font-size: 16px;
     border: none;
@@ -45,6 +45,8 @@
   }
   
   button:hover {
-    background-color: #0056b3;
+    background-color: rgb(144, 182, 223);
   }
+
+
