+++ thefesta/src/board/css/boardModify.css
@@ -29,9 +29,9 @@
   }
   
   .submitButton {
-    background-color: #007BFF;
-    color: #fff;
+    background-color: #ffffff;
+    color: #000000;
+    border: 1px solid #d9d9d9;
     padding: 10px;
-    border: none;
     cursor: pointer;
   }
